+++ utils/get-config-defaults.js
@@ -3,10 +3,11 @@
 // Modules
 const _ = require('lodash');
 const fs = require('fs');
+const { exit } = require('process');
 
 module.exports = options => {
   // Get the viaconf
-  if (options.webserver.type === 'nginx') options.defaultFiles.vhosts = 'default.conf.tpl';
+  if (_.startsWith(options.via, 'nginx')) options.defaultFiles.vhosts = 'default.conf.tpl';
 
   // Get the default db conf
   const dbConfig = _.get(options, 'database', 'mysql');
@@ -26,4 +27,7 @@ module.exports = options => {
       delete options.defaultFiles[type];
     }
   });
+
+  // Return
+  return options.defaultFiles;
 };